requirements.txt
openai>=1.0.0
langchain>=0.1.0
python-dotenv
tqdm
pydantic>=2.0
google.generativeai